refactor: lineService의 stations, lines 제거

+++ src/main/resources/schema.sql
@@ -3,7 +3,7 @@ create table if not exists STATION
     id bigint auto_increment not null,
     name varchar(255) not null,
     next_station bigint,
-    distance bigint,
+    distance int,
     line_id bigint,
     primary key(id)
 );
